split test_receive into test_symantec and test_urlvoid
